Update config with bug fixes

+++ setup.sh
@@ -1,4 +1,5 @@
 #!/bin/bash
+set -ex
 
 # Install Rust
 echo "(1 / 6) Installing common utilities"
@@ -6,13 +7,11 @@ echo ">>> Installing Rust..."
 curl --proto '=https' --tlsv1.2 -sSf https://sh.rustup.rs | sh
 . "$HOME/.cargo/env"
 
-# Node, nvm, and bun
 echo ">>> Installing Node, nvm, pnpm, and bun..."
 curl -o- https://raw.githubusercontent.com/nvm-sh/nvm/v0.39.7/install.sh | bash
 source ~/.bashrc
 nvm install 20
 curl -fsSL https://get.pnpm.io/install.sh | sh -
-curl -fsSL https://bun.sh/install | bash
 
 echo ">>> Installing Kitty..."
 curl -L https://sw.kovidgoyal.net/kitty/installer.sh | sh /dev/stdin
@@ -20,18 +19,10 @@ curl -L https://sw.kovidgoyal.net/kitty/installer.sh | sh /dev/stdin
 echo ">>> Installing Starship..."
 curl -sS https://starship.rs/install.sh | sh
 
-echo ">>> Installing Oh-My-Zsh..."
-sh -c "$(curl -fsSL https://raw.githubusercontent.com/ohmyzsh/ohmyzsh/master/tools/install.sh)"
-
 echo ">>> Installing zoxide"
 curl -sSfL https://raw.githubusercontent.com/ajeetdsouza/zoxide/main/install.sh | sh
 export PATH="$PATH:~/.local/bin"
 
-echo ">>> Installing neovim"
-curl -LO https://github.com/neovim/neovim/releases/latest/download/nvim-linux64.tar.gz
-sudo rm -rf /opt/nvim
-sudo tar -C /opt -xzf nvim-linux64.tar.gz
-export PATH="$PATH:/opt/nvim-linux64/bin"
 
 if [[ "$(uname -s)" == "Darwin" ]]; then
     # macOS
@@ -43,6 +34,14 @@ elif [[ -f /etc/redhat-release ]]; then
     echo "(3 / 6) Detected Red Hat-based system, installing software..."
     sudo dnf install zsh vim stow fish python3-neovim cmake expat-devel fontconfig-devel libxcb-devel freetype-devel libxml2-devel harfbuzz ripgrep fzf
     cargo install silicon
+
+    echo ">>> Installing neovim"
+    curl -LO https://github.com/neovim/neovim/releases/latest/download/nvim-linux64.tar.gz
+    sudo rm -rf /opt/nvim
+    sudo tar -C /opt -xzf nvim-linux64.tar.gz
+    export PATH="$PATH:/opt/nvim-linux64/bin"
+    echo 'export PATH="$PATH:/opt/nvim-linux64/bin"' >> ~/.zshrc
+    echo 'export PATH="$PATH:/opt/nvim-linux64/bin"' >> ~/.config/fish/config.fish
 
 elif [[ -f /etc/debian_version ]]; then
     # Debian
@@ -67,25 +66,32 @@ elif [[ -f /etc/debian_version ]]; then
     sudo apt-add-repository ppa:fish-shell/release-3
     sudo apt update
     sudo apt install -y fish
+
+    echo ">>> Installing neovim"
+    curl -LO https://github.com/neovim/neovim/releases/latest/download/nvim-linux64.tar.gz
+    sudo rm -rf /opt/nvim
+    sudo tar -C /opt -xzf nvim-linux64.tar.gz
+    export PATH="$PATH:/opt/nvim-linux64/bin"
+    echo 'export PATH="$PATH:/opt/nvim-linux64/bin"' >> ~/.zshrc
+    echo 'export PATH="$PATH:/opt/nvim-linux64/bin"' >> ~/.config/fish/config.fish
 fi
 
 echo ">>> Installing Poetry..."
 curl -sSL https://install.python-poetry.org | python3 -
 
-# Set up NvChad
-echo "(5 / 6) Setting up NvChad..."
-git clone https://github.com/NvChad/NvChad ~/.config/nvim --depth 1
-
 # Set up dotfiles
-echo "(6 / 6) Setting up dotfiles..."
-mv ~/.zshrc ~/.zshrc.bak
-rm ~/.config/nvim/.stylua.toml
+echo "(5 / 5) Setting up dotfiles..."
+
+if [ -f ~/.zshrc ]; then
+    mv ~/.zshrc ~/.zshrc.bak
+fi
+
+if [ -d ~/.config ]; then
+    mv ~/.config ~/.config.bak
+fi
+
 stow .
 
-if [ "$(uname)" = "Linux" ]; then
-    echo 'export PATH="$PATH:/opt/nvim-linux64/bin"' >> ~/.zshrc
-    echo 'export PATH="$PATH:/opt/nvim-linux64/bin"' >> ~/.config/fish/config.fish
-fi
 echo 'export PATH="$PATH:~/.local/bin"' >> ~/.zshrc
 
 echo "\n\n===================\nDone! Please restart your terminal.\n===================="
